datetime: Use common view for configuration
